mockly-backend/workflow/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-dotenv>=1.0
anthropic>=0.34
deepgram-sdk>=4.5
